# backend/chegirma_api_gateway/api/docs/swagger.yaml
definitions:
...
  ek_setting_service.CreateUpdateCategorySwag:
    properties:
      code:
        type: integer
      external_id:
        type: integer
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    required:
    - code
    - name
    - ru_name
    type: object
...
  ek_setting_service.CreateUpdateProductSwag:
    properties:
      category_id:
        type: string
      description:
        type: string
      discount_amount:
        type: integer
      from_time:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      owner_id:
        type: string
      price_after:
        type: integer
      price_before:
        type: integer
      to_time:
        type: string
      url:
        type: string
    type: object
...
  ek_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  ek_variables.SuccessResponse:
    properties:
      success:
        example: true
        type: boolean
    type: object
...
  setting_service.Category:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
...
  setting_service.CreatedResponse:
    properties:
      id:
        type: string
    type: object
...
  setting_service.GetAllCategoriesResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/setting_service.Category'
        type: array
      count:
        type: integer
    type: object
...
  setting_service.GetAllProductsResponse:
    properties:
      count:
        type: integer
      products:
        items:
          $ref: '#/definitions/setting_service.Product'
        type: array
    type: object
...
  setting_service.Product:
    properties:
      category:
        $ref: '#/definitions/setting_service.Category'
        type: object
      description:
        type: string
      discount_amount:
        type: integer
      from_time:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/setting_service.User'
        type: object
      price_after:
        type: integer
      price_before:
        type: integer
      to_time:
        type: string
      url:
        type: string
    type: object
...
  setting_service.User:
    properties:
      id:
        type: string
      image_url:
        type: string
      link:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /v1/category:
    get:
      consumes:
      - application/json
      description: API for getting all categoryes
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: soato
        in: query
        name: soato
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting_service.GetAllCategoriesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Getting All categories
      tags:
      - category
    post:
      consumes:
      - application/json
      description: API for creating category
      parameters:
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/ek_setting_service.CreateUpdateCategorySwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/setting_service.CreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Create category
      tags:
      - category
  /v1/category/{category_id}:
    get:
      consumes:
      - application/json
      description: API for getting category
      parameters:
      - description: category_id
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting_service.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Get Category
      tags:
      - category
    put:
      consumes:
      - application/json
      description: API for updating category
      parameters:
      - description: category_id
        in: path
        name: category_id
        required: true
        type: string
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/ek_setting_service.CreateUpdateCategorySwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting_service.CreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Update category
      tags:
      - category
...
  /v1/product:
    get:
      consumes:
      - application/json
      description: API for getting all productes
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: category_id
        in: query
        name: category_id
        type: string
      - description: owner_id
        in: query
        name: owner_id
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting_service.GetAllProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Getting All products
      tags:
      - product
    post:
      consumes:
      - application/json
      description: API for creating product
      parameters:
      - description: Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/ek_setting_service.CreateUpdateProductSwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/setting_service.CreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Create product
      tags:
      - product
  /v1/product/{product_id}:
    delete:
      consumes:
      - application/json
      description: API for deleting product
      parameters:
      - description: product_id
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ek_variables.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Delete product
      tags:
      - product
    get:
      consumes:
      - application/json
      description: API for getting product
      parameters:
      - description: product_id
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting_service.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Get Product
      tags:
      - product
    put:
      consumes:
      - application/json
      description: API for updating product
      parameters:
      - description: product_id
        in: path
        name: product_id
        required: true
        type: string
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/ek_setting_service.CreateUpdateProductSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting_service.CreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Update product
      tags:
      - product
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"